+++ lib/src/exceptions/exceptions.dart
@@ -133,8 +133,8 @@ abstract class SoLoudCppException extends SoLoudException {
         return const SoLoudBufferStreamWithReleasedBufferTypeCannotBeSeekedCppException();
       case PlayerErrors.audioFormatNotSupported:
         return const SoLoudAudioFormatNotSupportedCppException();
-      case PlayerErrors.opusOggVorbisLibsNotFound:
-        return const SoLoudOpusOggVorbisLibsNotFoundCppException();
+      case PlayerErrors.xiphLibsNotFound:
+        return const SoLoudXiphLibsNotFoundCppException();
       case PlayerErrors.busIdNotFound:
         return const SoLoudBusIdNotFoundCppException();
     }