+++ src/player.cpp
@@ -260,8 +260,8 @@ const std::string Player::getErrorString(PlayerErrors errorCode) const
         return "error: buffer stream with released buffer type cannot be seeked!";
     case audioFormatNotSupported:
         return "error: audio format not supported!";
-    case opusOggVorbisLibsNotFound:
-      return "error: opus ogg vorbis libraries not found!";
+    case xiphLibsNotFound:
+      return "error: Xiph libraries not found!";
     case busIdNotFound:
       return "error: bus id not found!";
     }